Refactor client and server code for improved structure and functionality
- Simplified the RemoteStorage trait implementations in client.
- Simplified mount options by removing allow_root field and related logic.
- Updated Docker configurations for client and server, including Dockerfiles and build scripts.
- Enhanced Windows build scripts for client with Inno Setup integration.

+++ client/src/fuse/unix.rs
@@ -1063,7 +1063,7 @@ impl From<fs_model::Attributes> for FileAttr {
             mtime: value.mtime.into(),
             ctime: value.ctime.into(),
             kind: value.kind.into(),
-            perm: u16::from(value.kind) | u16::from(value.perm),
+            perm: u16::from(value.kind) | value.perm,
             nlink: value.nlink,
             uid: value.uid,
             gid: value.gid,
